Fixbugs && refactoring code for signup/login
