+++ app.js
@@ -7,7 +7,8 @@ var bodyParser = require('body-parser');
 
 var db = require('./db/db');
 var index = require('./routes/index');
-var users = require('./routes/users');
+var signup = require('./routes/signup');
+var login = require('./routes/Login');
 var admin = require('./routes/admin');
 
 var app = express();
@@ -26,7 +27,8 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', index);
-app.use('/users', users);
+app.use('/signup', signup);
+app.use('/login', login);
 app.use('/admin', admin);
 
 // catch 404 and forward to error handler
