+++ views/login.ejs
@@ -2,18 +2,13 @@
 <html>
   <head>
     <title>Login</title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
   <body>
-    <h1>Ceci est un login</h1>
-
+    <h1>Login</h1>
     <form action="/login" method="post">
-
       <input type="email" name="email" placeholder="Votre email" required>
       <input type="password" name="password" placeholder="Votre mot de passe" required>
-      <input type="submit" name="submit" value="S'inscrire">
-
+      <input type="submit" name="submit" value="Valider">
     </form>
-
   </body>
 </html>
