+++ views/signup.ejs
@@ -2,18 +2,14 @@
 <html>
   <head>
     <title>Signup</title>
-    <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
   <body>
-    <h1>Ceci est un signup</h1>
-
+    <h1>Signup</h1>
     <form action="/signup" method="post">
-
       <input type="text" name="username" placeholder="Votre identifiant" required>
       <input type="email" name="email" placeholder="Votre email" required>
       <input type="password" name="password" placeholder="Votre mot de passe" required>
       <input type="submit" name="submit" value="S'inscrire">
-
     </form>
   </body>
 </html>
